requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
bs4==0.0.1
dj-static==0.0.6
Django==4.1.4
django-bootstrap4==22.3
django-crispy-forms==1.14.0
django-stdimage==6.0.1
fontawesomefree==6.2.1
gunicorn==20.1.0
pdfkit==1.0.0
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyodbc==4.0.35
soupsieve==2.3.2.post1
sqlparse==0.4.3
static3==0.7.0
tomli==2.0.1
whitenoise==6.2.0
